backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf>=3.15.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
tiktoken>=0.5.0
openai>=1.0.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pinned for passlib compatibility
